Switch to interactive dmenu mode
The problem with the rofi-script mode is that the rofi menu does
not disappear when the pinentry menu pops, which makes it impossible
to type the password.

Using a full rofi-dmenu mode would not allow integration with other
rofi modis.

This solution is an intermediate hack which aims at providing the best
of both worlds:
- it starts as a rofi-script to list password files
- an the second rofi-script call it uses coproc to start the script in
  rofi-dmenu mode

The main disadantage is that it forbids switching to another modi once a
password is selected. However, once a users have selected a password it is
unlikely that they wich to switch modi.